implement ship movement

+++ src/MovingObject.cpp
@@ -12,12 +12,12 @@ void MovingObject::update(double delta)
     Object::update(delta);
     
     // calculate new X position
-    int x_position_offset = (int)this->speed * sin(this->direction) * delta;
-    int new_x_position = this->get_x_position() + x_position_offset;
+    double x_position_offset = (int)this->speed * sin(this->direction) * delta;
+    double new_x_position = this->get_x_position() + x_position_offset;
 
     // calculate new Y position
-    int y_position_offset = (int)this->speed * cos(this->direction) * delta;
-    int new_y_position = this->get_y_position() + y_position_offset;
+    double y_position_offset = (int)this->speed * cos(this->direction) * delta;
+    double new_y_position = this->get_y_position() + y_position_offset;
 
     // on leaving screen borders, teleport to other side
     if(this->wrap_around_display)
@@ -42,13 +42,13 @@ void MovingObject::draw(Display display)
     this->undraw(display, this->previous_x_position, this->previous_y_position);
 }
 
-void MovingObject::set_x_position(int position)
+void MovingObject::set_x_position(double position)
 {
     this->previous_x_position = this->get_x_position();
     Object::set_x_position(position);
 }
 
-void MovingObject::set_y_position(int position)
+void MovingObject::set_y_position(double position)
 {
     this->previous_y_position = this->get_y_position();
     Object::set_y_position(position);
