+++ src/Object.cpp
@@ -1,27 +1,27 @@
 #include "Object.h"
 
-Object::Object(int x_position, int y_position)
+Object::Object(double x_position, double y_position)
 {
     this->x_position = x_position;
     this->y_position = y_position;
 }
 
-void Object::set_x_position(int position)
+void Object::set_x_position(double position)
 {
     this->x_position = position;
 }
 
-int Object::get_x_position()
+double Object::get_x_position()
 {
     return this->x_position;
 }
 
-void Object::set_y_position(int position)
+void Object::set_y_position(double position)
 {
     this->y_position = position;
 }
 
-int Object::get_y_position()
+double Object::get_y_position()
 {
     return this->y_position;
 }
